requirements.txt
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0 
flask>=3.1.1